Optimized with exception and types

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -9,15 +9,13 @@ import javax.validation.constraints.NotNull;
 import javax.validation.constraints.PositiveOrZero;
 import javax.validation.constraints.Size;
 import java.time.LocalDate;
-import java.util.ArrayList;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 
 @Data
 @Builder
 public class Film {
-    private int id;
+    private Long id;
     @NotBlank(message = "Название не может быть пустым или состоять только из пробелов")
     private String name;
     @Size(max = 200, message = "Длина описания больше 200 символов")
@@ -27,9 +25,6 @@ public class Film {
     private LocalDate releaseDate;
     @PositiveOrZero(message = "Продолжительность должна быть больше или равно 0")
     private int duration;
-    private final Set<Integer> likesId = new HashSet<>();
+    private final Set<Long> likesId = new HashSet<>();
 
-    //public void addLikeByUser(int userId){
-       // likesId.add(userId);
-   // }
 }
